Improve testing, add a new test, fix some things (#36)
First estimate the transition density `p_T(0,v)` for fix `T` of the non-linear diffusion `P` by forward-simulating paths and then taking an empirical, marginal distribution at time `T`. Then, compare to an alternative estimation method, where for *each value of `v`* (where `v`s are simulated at random) compute the stochastic importance weight by simulating proposal bridge and then assigning weight that corrects for the discrepancy between the proposal and the target. The contribution from the path is averaged out and only the contribution from the transition density from  the starting point to end-point `v` is left and it also comes from the correct law, because of the importance correction step.